requirements.main.txt
python-escpos==3.1
python-socketio==5.11.3
websocket-client==1.8.0
simple-websocket==1.0.0
requests==2.32.3
aiohttp==3.10.2
argparse==1.4.0
uuid==1.30
appdirs==1.4.4
kivy==2.3.0
python-dotenv==1.0.1
platformdirs=4.2.2